pyproject.toml
[project]
name = "memoir"
version = "0.1.0"
description = "A modular platform for collecting life stories and creating personalized documents"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "dspy>=2.5",
    "google-generativeai>=0.8",
    "openai>=1.0",
    "fastapi>=0.109",
    "uvicorn>=0.27",
    "python-multipart>=0.0.6",
    "httpx>=0.26",
    "python-dotenv>=1.0",
    "tenacity>=8.2",
    "pyjwt>=2.8",
    "sentry-sdk[fastapi]>=1.40",
    "email-validator>=2.0",
    "boto3>=1.34",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"

